+++ app/Http/Controllers/Panel/CommentController.php
@@ -10,11 +10,14 @@ use App\Models\Course;
 use App\Models\Episode;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-comments');
+
         $parentComments = Comment::where('comment_id',null)->orderByDesc('id')->paginate(4);
         $childrenComments = Comment::where('comment_id','!=',null)->orderByDesc('id')->paginate(4);
         return view('panel.comments.index',compact(['parentComments','childrenComments']));
@@ -22,6 +25,8 @@ class CommentController extends Controller
 
     public function save(CreateCommentRequest $request, Comment $comment)
     {
+        Gate::authorize('save-comment');
+
         $model = match ($comment->commentable_type) {
             'App\Models\Course' => Course::where('id',$comment->commentable_id)->first(),
             'App\Models\Post' => Post::where('id',$comment->commentable_id)->first(),
@@ -39,16 +44,22 @@ class CommentController extends Controller
 
     public function edit(Comment $comment)
     {
+        Gate::authorize('edit-comments');
+
         return view('panel.comments.edit',compact('comment'));
     }
 
     public function reply(Comment $comment)
     {
+        Gate::authorize('reply-comments');
+
         return view('panel.comments.reply',compact('comment'));
     }
 
     public function update(UpdateCommentRequest $request, Comment $comment)
     {
+        Gate::authorize('update-comments');
+
         $comment->update(
             $request->validated()
         );
@@ -58,6 +69,8 @@ class CommentController extends Controller
 
     public function display(Request $request, Comment $comment)
     {
+        Gate::authorize('is-approved-comments');
+
         $comment->update([
             'is_approved' => ! $comment->is_approved
         ]);
@@ -67,6 +80,8 @@ class CommentController extends Controller
 
     public function destroy(Request $request, Comment $comment)
     {
+        Gate::authorize('delete-comments');
+
         $comment->delete();
         $request->session()->flash('status','نظر مورد نظر با موفقیت حذف شد !');
         return back();
