+++ app/Http/Controllers/Panel/CouponController.php
@@ -7,22 +7,29 @@ use App\Http\Requests\Panel\Coupon\CreateCouponRequest;
 use App\Http\Requests\Panel\Coupon\UpdateCouponRequest;
 use App\Models\Coupon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CouponController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-coupons');
+
         $coupons = Coupon::orderByDesc('id')->paginate(5);
         return view('panel.coupons.index',compact('coupons'));
     }
 
     public function create()
     {
+        Gate::authorize('create-coupon');
+
         return view('panel.coupons.create');
     }
 
     public function store(CreateCouponRequest $request)
     {
+        Gate::authorize('store-coupon');
+
         $data = $request->validated();
         Coupon::create(
             $data
@@ -39,12 +46,15 @@ class CouponController extends Controller
 
     public function edit(Coupon $coupon)
     {
+        Gate::authorize('edit-coupons');
 //        $this->authorize('view', $coupon);
         return view('panel.coupons.edit',compact('coupon'));
     }
 
     public function isApproved(Request $request, Coupon $coupon)
     {
+        Gate::authorize('is-approved-coupons');
+
         $coupon->update([
             'is_approved' => ! $coupon->is_approved
         ]);
@@ -54,6 +64,8 @@ class CouponController extends Controller
 
     public function update(UpdateCouponRequest $request, Coupon $coupon)
     {
+        Gate::authorize('update-coupons');
+
         $data = $request->validated();
 
         if (!$request->expiry_date){
@@ -70,6 +82,8 @@ class CouponController extends Controller
 
     public function destroy(Request $request, Coupon $coupon)
     {
+        Gate::authorize('delete-coupons');
+
         $coupon->delete();
         $request->session()->flash('status','کد تخفیف مورد نظر با موفقیت حذف شد !');
         return back();
