+++ app/Http/Controllers/Panel/PermissionController.php
@@ -7,18 +7,23 @@ use App\Http\Requests\Panel\Permission\CreatePermissionRequest;
 use App\Http\Requests\Panel\Permission\UpdatePermissionRequest;
 use App\Models\Permission;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
 
     public function index()
     {
+        Gate::authorize('view-permissions');
+
         $permissions = Permission::paginate(5);
         return view('panel.permissions.index',compact('permissions'));
     }
 
     public function store(CreatePermissionRequest $request)
     {
+        Gate::authorize('store-permission');
+
         Permission::create(
             $request->validated()
         );
@@ -28,11 +33,15 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        Gate::authorize('edit-permissions');
+
         return view('panel.permissions.edit',compact('permission'));
     }
 
     public function update(UpdatePermissionRequest $request, Permission $permission)
     {
+        Gate::authorize('update-permissions');
+
         $permission->update(
             $request->validated()
         );
@@ -42,6 +51,8 @@ class PermissionController extends Controller
 
     public function destroy(Request $request, Permission $permission)
     {
+        Gate::authorize('delete-permissions');
+
         $permission->delete();
         $request->session()->flash('status','دسترسی مورد نظر با موفقیت حذف شد !');
         return back();
